agentic-backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
langchain-openai==0.3.32
openai==1.102.0
chromadb==1.0.20
pandas==2.3.2
numpy==2.3.2
yfinance==0.2.65
beautifulsoup4==4.13.5
requests==2.32.5
python-dotenv==1.1.1
sqlalchemy==2.0.43
pydantic==2.11.7
fastmcp
langchain-mcp-adapters
langgraph
backtesting
matplotlib
backtrader